Fixed (hopefully) import_awon and added param to last
Had to modify the import_awon script because once one station's import
had finished, the "last" date was advanced in the database so the
second station had nothing to do! (Duh.) So added a stnid param to
t411s#last, added that to the script, and created a test for it. Which
passes.